Separate localize and localize-mixin files (mimic intended intl file) (#4911)
